rename db files
