+++ dev/app/src/db/disease.py
@@ -1,6 +1,6 @@
 from psycopg2.extras import execute_batch  # type: ignore
 
-from db.get_connection import get_connection
+from dev.app.src.db.connection import get_connection
 from db.truncate_table import truncate_table
